Use i64 vs i32 in diesel for aggregate quantities
Source: https://github.com/msupply-foundation/open-msupply/commit/4d6bfa68182c7b5c1310abd214ca74f1ef28e9ec

+++ server/repository/src/db_diesel/consumption.rs
@@ -12,7 +12,7 @@ table! {
         id -> Text,
         item_id -> Text,
         store_id -> Text,
-        quantity -> Integer,
+        quantity -> BigInt,
         date -> Date,
     }
 }
@@ -25,7 +25,7 @@ pub struct ConsumptionRow {
     pub id: String,
     pub item_id: String,
     pub store_id: String,
-    pub quantity: i32,
+    pub quantity: i64,
     pub date: NaiveDate,
 }
